VersionControlSystem: Extract inner class
Making 'WorkingTree' a top level class makes both files
more readably due to size reduction and the level of
nesting. Both classes become more decoupled / separately
unit testable and the dependencies of 'WorkingTree' become
much more obvious.
