+++ downloader/src/main/kotlin/vcs/Cvs.kt
@@ -23,6 +23,7 @@ import ch.frankel.slf4k.*
 
 import com.here.ort.downloader.DownloadException
 import com.here.ort.downloader.VersionControlSystem
+import com.here.ort.downloader.WorkingTree
 import com.here.ort.model.Package
 import com.here.ort.utils.CommandLineTool
 import com.here.ort.utils.ProcessCapture
@@ -59,7 +60,7 @@ class Cvs : VersionControlSystem(), CommandLineTool {
             }
 
     override fun getWorkingTree(vcsDirectory: File) =
-            object : WorkingTree(vcsDirectory) {
+            object : WorkingTree(vcsDirectory, type) {
                 private val cvsDirectory = File(workingDir, "CVS")
 
                 override fun isValid(): Boolean {
