+++ downloader/src/main/kotlin/vcs/GitRepo.kt
@@ -22,6 +22,7 @@ package com.here.ort.downloader.vcs
 import ch.frankel.slf4k.*
 
 import com.here.ort.downloader.DownloadException
+import com.here.ort.downloader.WorkingTree
 import com.here.ort.model.Package
 import com.here.ort.model.VcsInfo
 import com.here.ort.utils.OS
